requirements.txt
Django==1.6.5
GnuPGInterface==0.3.2
South==1.0
apt-xapian-index==0.44
argparse==1.2.1
chardet==2.0.1
command-not-found==0.2.44
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
django-angular==0.7.3
djangorestframework==2.3.14
language-selector==0.1
python-apt==0.8.3ubuntu7
python-debian==0.1.21ubuntu1
six==1.7.3
ufw==0.31.1-1
wsgiref==0.1.2
